replace flow-router with react-router
