+++ .meteor/packages
@@ -40,8 +40,6 @@ settinghead:auto-nprogress
 meteorhacks:inject-initial
 
 # app common
-kadira:flow-router-ssr
-zimme:active-route@=2.3.3-rc.2
 universe:i18n
 momentjs:moment
 email
@@ -50,7 +48,6 @@ check
 
 # tests
 practicalmeteor:mocha
-# kadira:debug
 
 # performance
 meteorhacks:kadira
