+++ client/main.js
@@ -1,9 +1,14 @@
 import React from 'react';
 import '../imports/startup/client';
+import { Meteor } from 'meteor/meteor';
+import { render } from 'react-dom';
+import { renderRoutes } from '../imports/startup/client/routes.js';
 
-Meteor.startup(function () {
-    // initialize smooth scroll library
-    smoothScroll.init();
-    $('body').addClass('page');
-    $(".loading").remove();
+Meteor.startup(() => {
+  // initialize smooth scroll library
+  smoothScroll.init();
+  $('body').addClass('page');
+  $(".loading").remove();
+
+  render(renderRoutes(), document.getElementById('react-root'));
 });
