+++ imports/ui/components/navigation/SideBar.jsx
@@ -1,6 +1,5 @@
 import React, { Component, PropTypes } from 'react';
 import { Meteor } from 'meteor/meteor';
-import { FlowRouter } from 'meteor/kadira:flow-router-ssr';
 import classNames from 'classnames';
 
 export default class SideBar extends Component {
