+++ imports/ui/pages/people/PeopleContainer.jsx
@@ -1,18 +1,13 @@
 import React, { Component, PropTypes } from 'react';
 import { Meteor } from 'meteor/meteor';
 import { createContainer } from 'meteor/react-meteor-data';
-import { FlowRouter } from 'meteor/kadira:flow-router-ssr';
 import i18n from 'meteor/universe:i18n';
 const T = i18n.createComponent();
 
 class PeoplePage extends Component {
-  componentDidMount() {
-  }
-  componentWillUnmount() {
-  }
   gotoPerson(e) {
     e.preventDefault();
-    FlowRouter.go('/people/1');
+    // !TODO: go to person
   }
   render() {
     var customImage = {
