+++ package.json
@@ -13,6 +13,7 @@
     "react-addons-pure-render-mixin": "^15.4.1",
     "react-dom": "^15.4.2",
     "react-mounter": "^1.2.0",
+    "react-router": "^3.0.2",
     "tracker-component": "^1.3.21"
   }
 }
